+++ JsonToolsNppPlugin/Properties/AssemblyInfo.cs
@@ -28,5 +28,5 @@ using System.Runtime.InteropServices;
 //      Build Number
 //      Revision
 //
-[assembly: AssemblyVersion("8.2.0.1")]
-[assembly: AssemblyFileVersion("8.2.0.1")]
+[assembly: AssemblyVersion("8.2.0.2")]
+[assembly: AssemblyFileVersion("8.2.0.2")]